@@ -1,6 +1,7 @@
 package config.annotations.jsr330;
 
-import javax.inject.Named;
+
+import jakarta.inject.Named;
 
 @Named("clientDao1")
 public class ClientDaoImpl1 implements ClientDao{
