@@ -1,6 +1,6 @@
 package config.annotations.jsr330;
 
-import javax.inject.Named;
+import jakarta.inject.Named;
 
 @Named("clientDao2")
 public class ClientDaoImpl2 implements ClientDao {
